Implement detection and recording of local variables
This detects local variables as part of font-locking, and records them
to a cider-locals text property. It also detects regions that shouldn't
get dynamic font-lock (currently only the `ns` form) and applies the
cider-block-dynamic-font-lock property with value t.

Note this is a bit of a hack, and has known false positives, but a
proper solution would involve implementing a reader of Clojure code.

Known issues:

1. False positive for multi-arity functions. The args of all possible
   arities are gathered and applied to all scopes in the function.
2. False positives when a destructuting map has an :or clause.